+++ source/common/http/http1/conn_pool_legacy.h
@@ -1,14 +1,27 @@
 #pragma once
 
+#include <cstdint>
+#include <list>
+#include <memory>
+
+#include "envoy/event/deferred_deletable.h"
 #include "envoy/event/timer.h"
 #include "envoy/http/codec.h"
+#include "envoy/http/conn_pool.h"
+#include "envoy/network/connection.h"
+#include "envoy/stats/timespan.h"
 #include "envoy/upstream/upstream.h"
 
+#include "common/common/linked_object.h"
+#include "common/http/codec_client.h"
 #include "common/http/codec_wrappers.h"
-#include "common/http/conn_pool_base.h"
+#include "common/http/conn_pool_base_legacy.h"
+
+#include "absl/types/optional.h"
 
 namespace Envoy {
 namespace Http {
+namespace Legacy {
 namespace Http1 {
 
 /**
@@ -17,7 +30,7 @@ namespace Http1 {
  *       address. Higher layer code should handle resolving DNS on error and creating a new pool
  *       bound to a different IP address.
  */
-class ConnPoolImpl : public ConnPoolImplBase {
+class ConnPoolImpl : public ConnectionPool::Instance, public Legacy::ConnPoolImplBase {
 public:
   ConnPoolImpl(Event::Dispatcher& dispatcher, Upstream::HostConstSharedPtr host,
                Upstream::ResourcePriority priority,
@@ -28,9 +41,15 @@ public:
 
   // ConnectionPool::Instance
   Http::Protocol protocol() const override { return Http::Protocol::Http11; }
+  void addDrainedCallback(DrainedCb cb) override;
+  void drainConnections() override;
+  bool hasActiveConnections() const override;
+  ConnectionPool::Cancellable* newStream(StreamDecoder& response_decoder,
+                                         ConnectionPool::Callbacks& callbacks) override;
+  Upstream::HostDescriptionConstSharedPtr host() const override { return host_; };
 
   // ConnPoolImplBase
-  ActiveClientPtr instantiateActiveClient() override;
+  void checkForDrained() override;
 
 protected:
   struct ActiveClient;
@@ -51,7 +70,7 @@ protected:
 
     // Http::StreamCallbacks
     void onResetStream(StreamResetReason, absl::string_view) override {
-      parent_.parent().onDownstreamReset(parent_);
+      parent_.parent_.onDownstreamReset(parent_);
     }
     void onAboveWriteBufferHighWatermark() override {}
     void onBelowWriteBufferLowWatermark() override {}
@@ -64,24 +83,47 @@ protected:
 
   using StreamWrapperPtr = std::unique_ptr<StreamWrapper>;
 
-  struct ActiveClient : public ConnPoolImplBase::ActiveClient {
+  struct ActiveClient : ConnPoolImplBase::ActiveClient,
+                        LinkedObject<ActiveClient>,
+                        public Network::ConnectionCallbacks,
+                        public Event::DeferredDeletable {
     ActiveClient(ConnPoolImpl& parent);
+    ~ActiveClient() override;
 
-    ConnPoolImpl& parent() { return static_cast<ConnPoolImpl&>(parent_); }
+    void onConnectTimeout() override;
 
-    // ConnPoolImplBase::ActiveClient
-    bool hasActiveRequests() const override;
-    bool closingWithIncompleteRequest() const override;
-    StreamEncoder& newStreamEncoder(StreamDecoder& response_decoder) override;
+    // Network::ConnectionCallbacks
+    void onEvent(Network::ConnectionEvent event) override {
+      parent_.onConnectionEvent(*this, event);
+    }
+    void onAboveWriteBufferHighWatermark() override {}
+    void onBelowWriteBufferLowWatermark() override {}
 
+    ConnPoolImpl& parent_;
+    CodecClientPtr codec_client_;
+    Upstream::HostDescriptionConstSharedPtr real_host_description_;
     StreamWrapperPtr stream_wrapper_;
+    uint64_t remaining_requests_;
   };
 
+  using ActiveClientPtr = std::unique_ptr<ActiveClient>;
+
+  void attachRequestToClient(ActiveClient& client, StreamDecoder& response_decoder,
+                             ConnectionPool::Callbacks& callbacks);
+  virtual CodecClientPtr createCodecClient(Upstream::Host::CreateConnectionData& data) PURE;
+  void createNewConnection();
+  void onConnectionEvent(ActiveClient& client, Network::ConnectionEvent event);
   void onDownstreamReset(ActiveClient& client);
   void onResponseComplete(ActiveClient& client);
-  ActiveClient& firstReady() const { return static_cast<ActiveClient&>(*ready_clients_.front()); }
-  ActiveClient& firstBusy() const { return static_cast<ActiveClient&>(*busy_clients_.front()); }
+  void onUpstreamReady();
+  void processIdleClient(ActiveClient& client, bool delay);
 
+  Event::Dispatcher& dispatcher_;
+  std::list<ActiveClientPtr> ready_clients_;
+  std::list<ActiveClientPtr> busy_clients_;
+  std::list<DrainedCb> drained_callbacks_;
+  const Network::ConnectionSocket::OptionsSharedPtr socket_options_;
+  const Network::TransportSocketOptionsSharedPtr transport_socket_options_;
   Event::TimerPtr upstream_ready_timer_;
   bool upstream_ready_enabled_{false};
 };
@@ -101,12 +143,7 @@ public:
   CodecClientPtr createCodecClient(Upstream::Host::CreateConnectionData& data) override;
 };
 
-ConnectionPool::InstancePtr
-allocateConnPool(Event::Dispatcher& dispatcher, Upstream::HostConstSharedPtr host,
-                 Upstream::ResourcePriority priority,
-                 const Network::ConnectionSocket::OptionsSharedPtr& options,
-                 const Network::TransportSocketOptionsSharedPtr& transport_socket_options);
-
 } // namespace Http1
+} // namespace Legacy
 } // namespace Http
 } // namespace Envoy