+++ docs/root/intro/version_history.rst
@@ -5,6 +5,14 @@ Version history
 ================
 * config: use type URL to select an extension whenever the config type URL (or its previous versions) uniquely identify a typed extension, see :ref:`extension configuration <config_overview_extension_configuration>`.
 * retry: added a retry predicate that :ref:`rejects hosts based on metadata. <envoy_api_field_route.RetryPolicy.retry_host_predicate>`
+* upstream: combined HTTP/1 and HTTP/2 connection pool code. This means that circuit breaker
+  limits for both requests and connections apply to both pool types. Also, HTTP/2 now has
+  the option to limit concurrent requests on a connection, and allow multiple draining
+  connections. The old behavior is deprecated, but can be used during the deprecation
+  period by disabling runtime feature "envoy.reloadable_features.new_http1_connection_pool_behavior" or
+  "envoy.reloadable_features.new_http2_connection_pool_behavior" and then re-configure your clusters or
+  restart Envoy. The behavior will not switch until the connection pools are recreated. The new
+  circuit breaker behavior is described :ref:`here <arch_overview_circuit_break>`.
 * upstream: changed load distribution algorithm when all priorities enter :ref:`panic mode<arch_overview_load_balancing_panic_threshold>`. 
 
 1.13.0 (January 20, 2020)